refactor(mapgen): bind artifact validation to contracts (#2086)
refactor(mapgen): bind artifact validation to contracts

Make artifact modules a closed schema-bound kind: Core owns structural admission and exact validator binding, while mod artifacts keep only local relational laws. Inline artifact-private schemas, move genuinely shared domain schema vocabulary to model/schemas, remove the deprecated typed-array shape alias, and enforce the kind through Habitat blueprint structure and Grit authority.

refactor(mapgen): seal artifact and trace execution authority

refactor(mapgen): seal current engine state authority

+++ packages/civ7-map-policy/src/resource-legality.ts
@@ -32,11 +32,11 @@ export type ResourceLegalitySurface = {
   readonly width: number;
   readonly height: number;
   /** Engine-format biome index per tile. */
-  readonly biomeType: Uint8Array;
+  readonly biomeType: Int32Array;
   /** Engine-format terrain index per tile. */
-  readonly terrainType: Uint8Array;
+  readonly terrainType: Int32Array;
   /** Engine-format feature index per tile (-1 = none). */
-  readonly featureType: Int16Array;
+  readonly featureType: Int32Array;
   /** Engine water reading per tile (1 = water), used for adjacent-to-land flags. */
   readonly engineWaterMask: Uint8Array;
 };
